Chaitanya_P1/Selection-sheet-generator/selection_sheet_gen/requirements.txt
# ==========================================================
# Core & Configuration
# ==========================================================
python-dotenv==1.0.1
pydantic==2.9.2

# ==========================================================
# Google Cloud & AI
# ==========================================================
google-cloud-storage==3.4.1
google-generativeai==0.5.4
google-api-core==2.26.0
google-auth==2.41.1

# ==========================================================
# Database
# ==========================================================
psycopg2-binary==2.9.10

# ==========================================================
# PDF & OCR Processing
# ==========================================================
PyPDF2==3.0.1
pdfplumber==0.11.4
pdf2image==1.17.0
pytesseract==0.3.13  # Optional, for OCR support

# ==========================================================
# Template & Data Utilities
# ==========================================================
Jinja2==3.1.4
pandas==2.2.3
numpy==1.26.4

# ==========================================================
# Dev / QA Tools (optional)
# ==========================================================
pytest==8.3.3
black==24.8.0
flake8==7.1.1
